requirements.txt
numpy>=1.26
pandas>=2.0
pydantic>=2.7
pyyaml>=6
